add mvc config

+++ spring_aop_orm_transaction_cache/src/main/java/com/klasevich/spring/repository/EmployeeRepository.java
@@ -9,7 +9,7 @@ public interface EmployeeRepository {
 
     void saveEmployee(Employee employee);
 
-    Employee getEmployee(int id);
+    Employee getEmployee(Long id);
 
-    void deleteEmployee(int id);
+    void deleteEmployee(Long id);
 }
