+++ spring_aop_orm_transaction_cache/src/main/java/com/klasevich/spring/service/EmployeeService.java
@@ -10,7 +10,7 @@ public interface EmployeeService {
 
     void saveEmployee(Employee employee);
 
-    Employee getEmployee(int id);
+    Employee getEmployee(Long id);
 
-    void deleteEmployee(int id);
+    void deleteEmployee(Long id);
 }
